Couple of cleanups and fixes
1. Use volatile reads and writes to memory addresses. This one is
perhaps the most "could be done differently" but this approach seems
okay.
2. Remove [[nodiscard]] markers from set functions which return void, this
cleans up some compiler warnings.
3. Elide register reading in set when masked portion will result in the entire
register being replaced in the write. Without this you get a spurrious
read because the compiler cannot itself know that the masked read isn't
required for it's side effects.
4. Removed the specialized enum set logic for non-memory fields. The
implementation internally was identical since the code to use the
underlying type for reads isn't used with the pointer logic removed.